hide posts on scroll

+++ src/components/Home.vue
@@ -7,7 +7,7 @@
     <TopBar ref="topbar" subreddit="Popular" @params_changed="params_changed" />
     <ul class="list-group border-0 pt-0 mt-3">
         <Post v-for="post in posts" :post="post.data" :hidden="globalHiddenPosts.includes(post.data.permalink)"
-            @hide_post="hide_post" />
+            @hide_post="hide_post" class="post-element" :data-permalink="post.data.permalink" />
     </ul>
     <div v-if="!scroll_loaded" class="progress " role="progressbar" aria-label="Basic example" aria-valuenow="0"
         aria-valuemin="0" aria-valuemax="100">
@@ -77,13 +77,7 @@ async function params_changed() {
     get_posts();
 }
 
-function scroll_handle(el) {
-    if (el.target.scrollTop + el.target.clientHeight >= el.target.scrollHeight - window.innerWidth && scroll_loaded.value && after.value) {
-        scroll();
-    }
-}
-
-async function hide_post(postPermalink) {
+async function hide_post(postPermalink, force) {
     // this may cause performance issues, need to test on a daily basis and then move this to a parent container
     let hiddenPosts = JSON.parse(localStorage.getItem("hidden_posts"));
     // init hiddenPosts as an empty array if it doesn't exist
@@ -95,13 +89,34 @@ async function hide_post(postPermalink) {
         hiddenPosts.push(postPermalink);
         localStorage.setItem("hidden_posts", JSON.stringify(hiddenPosts));
         // hide post
-    } else {
+        globalHiddenPosts.value = hiddenPosts;
+    } else if (!force) {
         // remove post from hiddenPosts array
         hiddenPosts = hiddenPosts.filter((post) => post !== postPermalink);
         localStorage.setItem("hidden_posts", JSON.stringify(hiddenPosts));
         // unhide post
+        globalHiddenPosts.value = hiddenPosts;
     }
-    globalHiddenPosts.value = hiddenPosts;
+}
+
+async function handlePostsInViewport() {
+    const postElements = document.querySelectorAll('.post-element');
+
+    postElements.forEach((postElement, index) => {
+        const rect = postElement.getBoundingClientRect();
+        if (rect.bottom < 0) {
+            // The post element is above the viewport
+            // Add your logic here
+            hide_post(postElement.dataset.permalink, true);
+        }
+    });
+}
+
+function scroll_handle(el) {
+    handlePostsInViewport();
+    if (el.target.scrollTop + el.target.clientHeight >= el.target.scrollHeight - window.innerWidth && scroll_loaded.value && after.value) {
+        scroll();
+    }
 }
 
 onBeforeMount(() => {
